Corrected the deps2d() function and added tests
From what I understand, unit tests are meant to be in the
git root and not in the module itself. So, instead of putting
the matmodlab unit tests in the '/matmodlab/tests' directory, I
created '/tests' and put some unit tests there.

The vast majority of the tests are for the deps2d function. There
are two issues:

1) the 'dt' argument in deps2d() is not needed. I did not go through
   the codebase to fix it. It is unused.
2) there are no unit tests for rate_of_matrix_function() but, for
   now, the extensive testing for deps2d should be considered
   sufficient for coverage.