Fixed Tresca multi-step test.
Part of the problem was a clipping function on the lode parameter
that ensured that it never got to +1,-1. I added a few more 9's to
the 0.999,-0.999 and the behavior was much better.

Also, compare each column with the others instead of using np.allclose().
Do not use np.allclose() because small errors (~machine precision) are
introduced when dealing with large stresses (~1.0e6) and are propagated to
when the stress approaches zero.  np.allclose() has no memory of the
'average magnitude' of the stress over the test, and subsequently flags
what we know to be small deviations as errors.