Adding back add_step functionality.
The add_step method was taken away in favor of run_step so that steps
would be run as they were defined.  I added back the add_step method
that, just adds steps to a list of accumulated steps.  The steps are not
run until the `run` method is called.